+++ src/Cropperjs/Tests/Kernel/TwigAppKernel.php
@@ -14,7 +14,9 @@ namespace Symfony\UX\Cropperjs\Tests\Kernel;
 use Symfony\Bundle\FrameworkBundle\FrameworkBundle;
 use Symfony\Bundle\TwigBundle\TwigBundle;
 use Symfony\Component\Config\Loader\LoaderInterface;
+use Symfony\Component\DependencyInjection\Alias;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\Form\FormFactoryInterface;
 use Symfony\Component\HttpKernel\Kernel;
 use Symfony\UX\Cropperjs\CropperjsBundle;
 
@@ -27,7 +29,7 @@ class TwigAppKernel extends Kernel
 {
     use AppKernelTrait;
 
-    public function registerBundles()
+    public function registerBundles(): iterable
     {
         return [new FrameworkBundle(), new TwigBundle(), new CropperjsBundle()];
     }
@@ -37,6 +39,9 @@ class TwigAppKernel extends Kernel
         $loader->load(function (ContainerBuilder $container) {
             $container->loadFromExtension('framework', ['secret' => '$ecret', 'test' => true]);
             $container->loadFromExtension('twig', ['default_path' => __DIR__.'/templates', 'strict_variables' => true, 'exception_controller' => null]);
+
+            // create a public alias - FormFactoryInterface is removed otherwise
+            $container->setAlias('public_form_factory', new Alias(FormFactoryInterface::class, true));
         });
     }
 }