+++ src/LazyImage/Tests/Kernel/FrameworkAppKernel.php
@@ -26,7 +26,7 @@ class FrameworkAppKernel extends Kernel
 {
     use AppKernelTrait;
 
-    public function registerBundles()
+    public function registerBundles(): iterable
     {
         return [new FrameworkBundle(), new LazyImageBundle()];
     }