+++ src/LazyImage/Tests/Kernel/TwigAppKernel.php
@@ -27,7 +27,7 @@ class TwigAppKernel extends Kernel
 {
     use AppKernelTrait;
 
-    public function registerBundles()
+    public function registerBundles(): iterable
     {
         return [new FrameworkBundle(), new TwigBundle(), new LazyImageBundle()];
     }