+++ src/Turbo/Bridge/Mercure/Resources/assets/src/turbo_stream_controller.js
@@ -7,7 +7,7 @@
  * file that was distributed with this source code.
  */
 
-import { Controller } from 'stimulus';
+import { Controller } from '@hotwired/stimulus';
 import { connectStreamSource, disconnectStreamSource } from '@hotwired/turbo';
 
 /**