+++ src/Turbo/Resources/assets/test/turbo_controller.test.js
@@ -9,7 +9,7 @@
 
 'use strict';
 
-import { Application } from 'stimulus';
+import { Application } from '@hotwired/stimulus';
 import { getByTestId } from '@testing-library/dom';
 import { clearDOM, mountDOM } from '@symfony/stimulus-testing';
 import TurboController from '../src/turbo_controller';