+++ src/Chartjs/CHANGELOG.md
@@ -1,7 +1,10 @@
 # CHANGELOG
 
-## 1.4
+## 2.0
 
+-   Support for `stimulus` version 2 was removed and support for `@hotwired/stimulus`
+    version 3 was added. See the [@symfony/stimulus-bridge CHANGELOG](https://github.com/symfony/stimulus-bridge/blob/main/CHANGELOG.md#300)
+    for more details.
 -   Support added for Symfony 6
 
 ## 1.3